sysuser: move external exports to index.js
